[ecommerce] (refactor/middleware): Refactor middleware &  applied Advance middleware
