+++ ecommerce/cmd/routes.go
@@ -27,8 +27,8 @@ func initRoutes(mux *http.ServeMux, manager *middleware.Manager) {
 	mux.Handle("GET /products", 
 		manager.With(
 			http.HandlerFunc(handlers.GetProducts),
-	),
-)
+		),
+	)
 
 	mux.Handle("POST /products", 
 		manager.With(
