+++ ecommerce/cmd/serve.go
@@ -4,25 +4,28 @@ import (
 	"fmt"
 	"net/http"
 
-	
-	"ecommerce/global_router"
-	"ecommerce/middleware"
-	
+
+	"ecommerce/middleware"	
 )
 
 func Serve(){
+	
+
 	manager := middleware.NewManager()
+	manager.Use(
+		middleware.Preflight,
+		middleware.Cors,
+		middleware.Logger,
+	)
 
-	manager.Use(middleware.Logger, middleware.Hudai)
+	mux := http.NewServeMux()	
+	wrappedMux := manager.WrapMux(mux)
 
-	mux := http.NewServeMux()
 
 	initRoutes(mux, manager)
 
-	globalRouter := global_router.GlobalRouter(mux)
-
 	fmt.Println("Server running on port :8080")
-	err := http.ListenAndServe(":8080", globalRouter) //" Failed to start the server"
+	err := http.ListenAndServe(":8080", wrappedMux) //" Failed to start the server"
 	if err != nil {
 		fmt.Println("Error starting the server", err)
 	}
