+++ ecommerce/middleware/cors.go
@@ -1,22 +1,15 @@
-package global_router 
+package middleware 
 
 import "net/http"
 
 
-func GlobalRouter(mux *http.ServeMux) http.Handler {
-	handleAllReq := func(w http.ResponseWriter, r *http.Request) {
+func Cors(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Access-Control-Allow-Origin", "*")
 		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE, OPTIONS")
 		w.Header().Set("Access-Control-Headers", "Content-Type, Habib")
 		w.Header().Set("Content-Type", "application/json")
 		
-		if r.Method == "OPTIONS" {
-			w.WriteHeader(200)
-			return 
-		} 
-		
-		mux.ServeHTTP(w, r)
-	}
-
-	return http.HandlerFunc(handleAllReq)
+		next.ServeHTTP(w, r)
+	})
 }
