+++ ecommerce/middleware/logger.go
@@ -1,4 +1,5 @@
 package middleware
+
 import (
 	"log"
 	"net/http"
