+++ ecommerce/middleware/preflight.go
@@ -1,13 +1,16 @@
 package middleware
 
 import (
-	"log"
 	"net/http"
 )
 
-func Hudai(next http.Handler) http.Handler {
+func Preflight(next http.Handler) http.Handler {
 	return http.HandlerFunc(func (w http.ResponseWriter, r *http.Request) {
-		log.Println("ami hudai middleware")
+		if r.Method == "OPTIONS" {
+			w.WriteHeader(200)
+			return  
+		}
+		
 		next.ServeHTTP(w, r)
 	})
 }
